# openapi/keeppage-api.yaml
openapi: 3.1.0
info:
  title: KeepPage API
  version: 0.1.0
  description: Contract source for the gradual Go backend migration.
servers:
  - url: http://127.0.0.1:8787
    description: TypeScript API
  - url: http://127.0.0.1:8788
    description: Go API migration slice
security:
  - bearerAuth: []
paths:
  /:
    get:
      operationId: getApiInfo
      responses:
        "200":
          description: API metadata.
          content:
            application/json:
              schema:
                type: object
                required: [name, storage]
                properties:
                  name:
                    type: string
                  storage:
                    $ref: "#/components/schemas/StorageDriver"
  /health:
    get:
      operationId: getHealth
      security: []
      responses:
        "200":
          description: Health status.
          content:
            application/json:
              schema:
                type: object
                required: [status, storage, uptimeSec, tables, now]
                properties:
                  status:
                    const: ok
                  storage:
                    $ref: "#/components/schemas/StorageDriver"
                  uptimeSec:
                    type: number
                    minimum: 0
                  tables:
                    type: array
                    items:
                      type: string
                  now:
                    type: string
                    format: date-time
  /bookmarks:
    get:
      operationId: listBookmarks
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: quality
          in: query
          schema:
            $ref: "#/components/schemas/QualityGrade"
        - name: view
          in: query
          schema:
            $ref: "#/components/schemas/BookmarkListView"
        - name: domain
          in: query
          schema:
            type: string
        - name: folderId
          in: query
          schema:
            type: string
        - name: tagId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated bookmarks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkSearchResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: createBookmark
      description: Go migration convenience endpoint. Uses the same payload and behavior as ingest bookmark creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBookmarkRequest"
      responses:
        "200":
          description: Existing bookmark merged or skipped.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBookmarkResponse"
        "201":
          description: Bookmark created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBookmarkResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /ingest/bookmarks:
    post:
      operationId: ingestBookmark
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBookmarkRequest"
      responses:
        "200":
          description: Existing bookmark merged or skipped.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBookmarkResponse"
        "201":
          description: Bookmark created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBookmarkResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-keeppage-api-key
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    StorageDriver:
      type: string
      enum: [memory, postgres]
    QualityGrade:
      type: string
      enum: [high, medium, low]
    BookmarkListView:
      type: string
      enum: [all, recent, favorites]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        details: {}
    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
    Folder:
      type: object
      required: [id, name, path]
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        parentId:
          type:
            - string
            - "null"
    Bookmark:
      type: object
      required:
        - id
        - sourceUrl
        - title
        - domain
        - note
        - isFavorite
        - tags
        - versionCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        sourceUrl:
          type: string
          format: uri
        canonicalUrl:
          type: string
          format: uri
        title:
          type: string
        domain:
          type: string
        faviconUrl:
          type: string
          format: uri
        coverImageUrl:
          type: string
          format: uri
        note:
          type: string
        isFavorite:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        folder:
          $ref: "#/components/schemas/Folder"
        latestVersionId:
          type: string
        versionCount:
          type: integer
          minimum: 0
        latestQuality: {}
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookmarkSearchResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
        total:
          type: integer
          minimum: 0
    IngestBookmarkRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
          maxLength: 500
        note:
          type: string
          maxLength: 4000
        tags:
          type: array
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 80
        folderPath:
          type: string
          minLength: 1
          maxLength: 500
        dedupeStrategy:
          type: string
          enum: [merge, skip]
          default: merge
    IngestBookmarkResponse:
      type: object
      required: [bookmarkId, status, deduplicated, bookmark]
      properties:
        bookmarkId:
          type: string
        status:
          type: string
          enum: [created, merged, skipped]
        deduplicated:
          type: boolean
        bookmark:
          $ref: "#/components/schemas/Bookmark"